+++ changelog.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Events trait with trigger method.
+- Repository layer method to get a specified layer from the repository
+  decorator stack.
 
 ## [0.11.8] - 2021-01-10
 ### Fixed